fix(ci): add missing __init__.py files and fix mypy errors
- Add __init__.py to 10 subpackages that were missing them (agents,
  api, data, execution, monitoring, orchestration, portfolio, risk,
  types, utils) — setuptools requires these for proper package discovery
- Fix .gitignore `data/` pattern to `/data/` so it doesn't ignore
  src/aiswarm/data/
- Fix 3 mypy no-any-return errors with explicit float() casts

+++ src/aiswarm/agents/strategy/momentum_agent.py
@@ -29,7 +29,7 @@ def _sma(candles: list[OHLCV], period: int) -> float | None:
     if len(candles) < period:
         return None
     closes = [c.close for c in candles[-period:]]
-    return sum(closes) / len(closes)
+    return float(sum(closes) / len(closes))
 
 
 def _trend_consistency(candles: list[OHLCV], period: int) -> float:
